fix(error-reporter): Restore CustomReporter to default rabbit config
