+++ async/async-rabbit/src/main/java/org/reactivecommons/async/rabbit/DynamicRegistryImp.java
@@ -4,7 +4,6 @@ package org.reactivecommons.async.rabbit;
 import lombok.RequiredArgsConstructor;
 import org.reactivecommons.async.api.DynamicRegistry;
 import org.reactivecommons.async.api.handlers.DomainEventHandler;
-import org.reactivecommons.async.api.handlers.EventHandler;
 import org.reactivecommons.async.api.handlers.QueryHandler;
 import org.reactivecommons.async.api.handlers.QueryHandlerDelegate;
 import org.reactivecommons.async.api.handlers.registered.RegisteredEventListener;
@@ -33,7 +32,8 @@ public class DynamicRegistryImp implements DynamicRegistry {
 
     @Override
     public <T, R> void serveQuery(String resource, QueryHandler<T, R> handler, Class<R> queryClass) {
-        resolver.addQueryHandler(new RegisteredQueryHandler<>(resource, (ignored, message) -> handler.handle(message), queryClass));
+        resolver.addQueryHandler(new RegisteredQueryHandler<>(resource, (ignored, message) -> handler.handle(message)
+                , queryClass));
     }
 
     @Override
